Creating an Account and receiving emails generically.
Class Korisnikis used to create a user, enter the name, surname and year of birth. Then an email is generated and it receives info about the user. In Class Account, an account is created for the user and added to the list, if the user already exists, it cannot be added to the list and I have a display of all users and emails in the list.