Chat creation and simulation of correspondence, with adding users and checking when adding.
First we create a user, each user has a name, username and password. I have a global registration of users who have to confirm a password and meet some conditions before they can be created. They have three types of companies: Public, Private and Group. They have controls for everything and what they need to fill in for the message to be added to the chat.